Split components rendering & logic
